HDDS-7214 Continuous start & stop can have hanging threads in stopping
